+++ templates/header.php
@@ -1,0 +1,19 @@
+        <header class="main-header">
+            <a href="#">
+                <img src="img/logo.png" width="153" height="42" alt="Логитип Дела в порядке">
+            </a>
+
+            <div class="main-header__side">
+                <a class="main-header__side-item button button--plus" href="#">Добавить задачу</a>
+                <div class="main-header__side-item user-menu">
+                    <div class="user-menu__image">
+                        <img src="img/user-pic.jpg" width="40" height="40" alt="Пользователь">
+                    </div>
+
+                    <div class="user-menu__data">
+                        <p>Константин</p>
+                        <a href="#">Выйти</a>
+                    </div>
+                </div>
+            </div>
+        </header>
